refactor: 优化项目结构和工具函数
- 将 `utils.ts` 重命名为 `tailwind-utils.ts` 以提高代码可读性
- 创建 `chatUtils.ts` 包含消息创建辅助函数
- 重构 `useChatHandlers` 中的消息生成逻辑
- 提高代码模块化和可复用性
- 更新组件中的导入路径
