+++ docs/CHANGELOG.md
@@ -12,6 +12,10 @@
 - 优化消息显示组件
   - 区分不同角色消息样式
   - 调整消息布局和间距
+- 优化用户体验
+  - 添加消息发送状态指示
+  - 优化错误提示展示
+  - 改进消息加载动画
 
 ### 🛠 技术改进
 - 重构消息处理逻辑
@@ -21,10 +25,15 @@
 - 完善类型定义
   - 区分前端显示消息和API请求消息类型
   - 添加必要的类型注解
+- 优化代码结构
+  - 抽取公共处理函数
+  - 规范化错误处理流程
+  - 改进组件复用性
 
 ### 📝 文档
 - 添加流式对话实现教程
 - 更新项目管理文档
+- 完善代码注释和文档说明
 
 ## [0.2.0-dev] - 2025-01-22
